backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
